# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PokerHub API
  version: 3.0.0
paths:
...
  /admin/kyc/{id}/denial:
    get:
      summary: Get KYC denial reason
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Denial reason
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycDenialResponse'
...
  /admin/audit-logs:
    get:
      summary: Get audit logs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          required: false
          schema:
            type: string
        - in: query
          name: type
          required: false
          schema:
            $ref: '#/components/schemas/AuditLogType'
        - in: query
          name: user
          required: false
          schema:
            type: string
        - in: query
          name: dateFrom
          required: false
          schema:
            type: string
        - in: query
          name: dateTo
          required: false
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogsResponse'
  /admin/audit/log-types:
    get:
      summary: Get audit log types
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Audit log types
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogTypesResponse'
  /admin/log-types:
    get:
      summary: Get audit log type classes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Audit log type classes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogTypeClasses'
  /admin/audit-logs/{id}/review:
    post:
      summary: Mark audit log reviewed
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated audit log entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogEntry'
  /admin/security-alerts:
    get:
      summary: Get security alerts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Security alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityAlertsResponse'
  /admin/security-alerts/{id}/ack:
    post:
      summary: Acknowledge a security alert
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Acknowledged security alert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertItem'
        '404':
          description: Security alert not found
  /admin/events:
    get:
      summary: Get admin events
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Admin events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventsResponse'
  /admin/events/{id}/ack:
    post:
      summary: Acknowledge an admin event
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Admin event acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Admin event not found
  /admin/tabs:
    get:
      summary: Get admin dashboard tabs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard tabs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTabResponse'
    post:
      summary: Create admin dashboard tab
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminTabRequest'
      responses:
        '200':
          description: Created admin tab
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTab'
  /admin/tabs/{id}:
    get:
      summary: Get admin tab metadata
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Admin tab metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTabMeta'
    patch:
      summary: Update admin dashboard tab
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminTabRequest'
      responses:
        '200':
          description: Updated admin tab
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTab'
    delete:
      summary: Delete admin dashboard tab
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Admin tab deleted
  /admin/revenue-breakdown:
    get:
      summary: Get revenue breakdown
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: range
          schema:
            type: string
            enum: [today, week, month, all]
      responses:
        '200':
          description: Revenue breakdown
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueBreakdown'
  /admin/wallet/reconcile/mismatches:
    get:
      summary: List wallet reconciliation mismatches
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wallet reconciliation mismatches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletReconcileMismatchesResponse'
  /admin/wallet/reconcile/mismatches/{account}/ack:
    post:
      summary: Acknowledge wallet reconciliation mismatch
      security:
        - bearerAuth: []
      parameters:
        - name: account
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mismatch acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletReconcileMismatchAcknowledgement'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    RevenueStream:
      type: object
      properties:
        label:
          type: string
        pct:
          type: number
        value:
          type: number
      required:
        - label
        - pct
    RevenueBreakdown:
      type: object
      properties:
        currency:
          type: string
          minLength: 3
          maxLength: 3
        streams:
          type: array
          items:
            $ref: '#/components/schemas/RevenueStream'
      required:
        - currency
        - streams
...
    AuditLogType:
      type: string
    AuditLogEntry:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        type:
          $ref: '#/components/schemas/AuditLogType'
        description:
          type: string
        user:
          type: string
        ip:
          type: string
        reviewed:
          type: boolean
        reviewedBy:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - timestamp
        - type
        - description
        - user
        - ip
        - reviewed
        - reviewedBy
        - reviewedAt
    AuditLogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
        total:
          type: integer
      required:
        - logs
        - total
    AuditLogTypesResponse:
      type: object
      properties:
        types:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogType'
      required:
        - types
    LogTypeClasses:
      type: object
      additionalProperties:
        type: string
...
    AlertItem:
      type: object
      properties:
        id:
          type: string
        severity:
          type: string
          enum: [danger, warning]
        title:
          type: string
        body:
          type: string
        time:
          type: string
        resolved:
          type: boolean
      required:
        - id
        - severity
        - title
        - body
        - time
        - resolved
    SecurityAlertsResponse:
      type: array
      items:
        $ref: '#/components/schemas/AlertItem'
...
    KycDenialResponse:
      type: object
      properties:
        accountId:
          type: string
        reason:
          type: string
          nullable: true
      required:
        - accountId
...
    WalletReconcileMismatch:
      type: object
      properties:
        account:
          type: string
        balance:
          type: number
        journal:
          type: number
        delta:
          type: number
        date:
          type: string
          format: date-time
      required:
        - account
        - balance
        - journal
        - delta
        - date
    WalletReconcileMismatchAcknowledgement:
      type: object
      properties:
        account:
          type: string
        acknowledgedBy:
          type: string
        acknowledgedAt:
          type: string
          format: date-time
      required:
        - account
        - acknowledgedBy
        - acknowledgedAt
    WalletReconcileMismatchesResponse:
      type: object
      properties:
        mismatches:
          type: array
          items:
            $ref: '#/components/schemas/WalletReconcileMismatch'
      required:
        - mismatches
...